Implement rootPathPrefix support for feign-ranger
Problem:
When a call using feign-ranger is done for `@RequestLine("GET /v1/test")`,
with the current implementation the call gets translated to:
`http://Address-from-ranger-for-backend-service/v1/test`
If we need to translate this call to a call accepted by gateway, it
should become:
`http://Address-from-ranger-for-gateway-service/apis/backend/v1/test`.

When the client calls have to pass through api gateways which expect
a prefix, each http call needs to have a prefix that will be intercepted
by the api gateway which processes the request based on the prefix and
calls the actual backend where this call is intended to go. It would be
better to have this capability in feign-ranger.

Fix:
Implement capability to add rootPathPrefix as a builder argument to
facilitate this functionality.

Signed-off-by: Pranith Kumar K <pranith.karampuri@phonepe.com>